界面选择服务器登陆功能开发；优化界面

+++ landlords-client-javafx/src/main/java/priv/zxw/ratel/landlords/client/javafx/ui/view/UIObject.java
@@ -37,7 +37,9 @@ public abstract class UIObject extends Stage {
 
         // 关闭netty
         NettyClient nettyClient = BeanUtil.getBean("nettyClient");
-        nettyClient.destroy();
+        if (nettyClient != null) {
+            nettyClient.destroy();
+        }
     }
 
     public <T> T $(String id, Class<T> clazz) {
